refactor(auth): application command와 result DTO 분리
