+++ backend/src/main/java/com/knot/backend/auth/application/MemberNicknameService.java
@@ -17,7 +17,7 @@ public class MemberNicknameService {
     private final OAuthIdentityService oauthIdentityService;
 
     @Transactional
-    public Member completeNickname(
+    public Member completeNicknameSetup(
             OAuthUser oauthUser,
             String nickname
     ) {
