+++ backend/src/main/java/com/knot/backend/auth/application/dto/command/CompleteNicknameCommand.java
@@ -1,0 +1,7 @@
+package com.knot.backend.auth.application.dto.command;
+
+public record CompleteNicknameCommand(
+        String nicknameToken,
+        String nickname
+) {
+}
